refactor: refactor editor related modules
