+++ packages/ui/src/controllers/ui/ui-desktop.controller.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Disposable, IUniverInstanceService, LifecycleService, LifecycleStages, toDisposable } from '@univerjs/core';
+import { Disposable, isInternalEditorID, IUniverInstanceService, LifecycleService, LifecycleStages, toDisposable } from '@univerjs/core';
 import { IRenderManagerService } from '@univerjs/engine-render';
 import type { IDisposable } from '@wendellhu/redi';
 import { Inject, Injector, Optional } from '@wendellhu/redi';
@@ -53,14 +53,12 @@ export class DesktopUIController extends Disposable {
                     this.disposeWithMe(this._layoutService.registerCanvasElement(canvasElement as HTMLCanvasElement));
                 }
 
-                // TODO: this is subject to change in the future
+                // TODO: this is subject to change in the future for Uni-mode
                 this._renderManagerService.currentRender$.subscribe((renderId) => {
                     if (renderId) {
                         const render = this._renderManagerService.getRenderById(renderId)!;
                         if (!render.unitId) return;
-
-                        // const unitType = this._instanceService.getUnitType(render.unitId);
-                        // if (unitType !== UniverInstanceType.UNIVER_SHEET) return;
+                        if (isInternalEditorID(render.unitId)) return;
 
                         render.engine.setContainer(canvasElement);
                     }
